Rename android tests package
